generator_script/requirements.txt
pandas~=2.2.3
python-dotenv~=1.1.0
boto3~=1.37.24